requirements.txt
# 基础依赖
Flask==3.0.2 
Werkzeug>=2.3.0
Jinja2>=3.1.2
itsdangerous>=2.1.2
click>=8.1.3
python-dotenv>=0.19.0
requests>=2.26.0
PyYAML>=5.4.1
aiohttp>=3.8.0
beautifulsoup4>=4.9.3
openai>=1.0.0
mcp
fastapi-mcp
# MCP相关依赖（需要从私有源安装或本地安装）
# mcp-server
# mcp-tools
# mcp-fastapi

# 测试依赖
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-asyncio>=0.18.0
pytest-mock>=3.6.1
httpx>=0.23.0
aioresponses>=0.7.1
